move GetCommitment() down into base & change base fields to protected
