+++ src/SFA.DAS.ProviderApprenticeshipsService.Web/Orchestrators/BaseCommitmentOrchestrator.cs
@@ -7,21 +7,44 @@ using MediatR;
 using SFA.DAS.Commitments.Api.Types;
 using SFA.DAS.Commitments.Api.Types.Commitment;
 using SFA.DAS.Commitments.Api.Types.Commitment.Types;
+using SFA.DAS.HashingService;
 using SFA.DAS.ProviderApprenticeshipsService.Application.Exceptions;
 using SFA.DAS.ProviderApprenticeshipsService.Application.Queries.GetCommitment;
+using SFA.DAS.ProviderApprenticeshipsService.Domain.Interfaces;
 
 namespace SFA.DAS.ProviderApprenticeshipsService.Web.Orchestrators
 {
     public class BaseCommitmentOrchestrator
     {
-        private readonly IMediator _mediator;
+        protected readonly IMediator _mediator;
+        protected readonly IHashingService _hashingService;
+        protected readonly IProviderCommitmentsLogger _logger;
 
-        public BaseCommitmentOrchestrator(IMediator mediator)
+        public BaseCommitmentOrchestrator(IMediator mediator,
+            IHashingService hashingService, IProviderCommitmentsLogger logger)
         {
-            if (mediator == null)
-                throw new ArgumentNullException(nameof(mediator));
+            // we keep null checks here, as this is a base class
+            _mediator = mediator ?? throw new ArgumentNullException(nameof(mediator));
+            _hashingService = hashingService ?? throw new ArgumentNullException(nameof(hashingService));
+            _logger = logger ?? throw new ArgumentNullException(nameof(logger));
+        }
 
-            _mediator = mediator;
+        public async Task<CommitmentView> GetCommitment(long providerId, string hashedCommitmentId)
+        {
+            return await GetCommitment(providerId, _hashingService.DecodeValue(hashedCommitmentId));
+        }
+
+        public async Task<CommitmentView> GetCommitment(long providerId, long commitmentId)
+        {
+            _logger.Info($"Getting commitment:{commitmentId} for provider:{providerId}", providerId, commitmentId);
+
+            var data = await _mediator.SendAsync(new GetCommitmentQueryRequest
+            {
+                ProviderId = providerId,
+                CommitmentId = commitmentId
+            });
+
+            return data.Commitment;
         }
 
         //todo: the consumers of this get the commitment themeselves,
